Fix data type
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@110 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/LuceneKit/Headers/Index/LCIndexReader.h
@@ -77,7 +77,7 @@ typedef enum _LCFieldOption
 - (void) setNorm: (int) doc field: (NSString *) field floatValue: (float) value;
 - (LCTermEnum *) terms;
 - (LCTermEnum *) termsWithTerm: (LCTerm *) t;
-- (int) docFreq: (LCTerm *) t;
+- (long) docFreq: (LCTerm *) t;
 - (id <LCTermDocs>) termDocsWithTerm: (LCTerm *) term;
 - (id <LCTermDocs>) termDocs;
 - (id <LCTermPositions>) termPositionsWithTerm: (LCTerm *) term;
